LINKER EXORCISM: Fixed 'undefined reference' to setCalibration by aligning unsigned long types

+++ kegstill_nuclear_final/valve.h
@@ -1,11 +1,11 @@
-// valve.h - 3-wire smart ball valve driver (internal limit switches, timed positioning)
+// valve.h - 3-wire smart ball valve driver
 #pragma once
 #include <Arduino.h>
 
 namespace valve {
   enum State : uint8_t {
-    ST_UNKNOWN     = 0,   // boot, never homed
-    ST_BOOT_HOMING = 1,   // driving closed to find home (timed)
+    ST_UNKNOWN     = 0,   
+    ST_BOOT_HOMING = 1,   
     ST_IDLE        = 2,
     ST_OPENING     = 3,
     ST_CLOSING     = 4,
@@ -14,20 +14,20 @@ namespace valve {
 
   struct Status {
     State    state;
-    float    position;        // 0-100, best estimate
-    uint8_t  target;          // 0-100
+    float    position;        
+    uint8_t  target;          
     bool     calibrated;
-    uint32_t openTimeMs;      // full-open travel duration (manually entered)
-    uint32_t closeTimeMs;     // full-close travel duration
+    unsigned long openTimeMs;      
+    unsigned long closeTimeMs;     
     String   faultReason;
   };
 
   void begin();
-  void poll();                                    // call every loop
+  void poll();                                    
   void setPosition(uint8_t pct);
   uint8_t getPosition();
-  void setCalibration(uint32_t openMs, uint32_t closeMs); // user-entered times
-  void rehome();                                  // re-drive to closed (re-zero)
+  void setCalibration(unsigned long openMs, unsigned long closeMs); 
+  void rehome();                                  
   void emergencyStop();
   void clearFault();
   Status getStatus();
